test: move CustomObjectId admin registration to setUp/tearDown
