+++ tests/test_app/tests/test_admin.py
@@ -573,15 +573,31 @@ class AdminRegisterInlineTest(TestBase):
         self.assertTrue(reversion.is_registered(TestModelGenericInline))
 
 
+class TestModelCustomObjectIdAdmin(VersionAdmin):
+
+    def reversion_register(self, model, **kwargs):
+        kwargs["object_id_field"] = "slug"
+        super().reversion_register(model, **kwargs)
+
+    def get_reversion_object_id(self, request, object_id):
+        obj = self.get_object(request, admin.utils.unquote(object_id))
+        return str(obj.slug) if obj else admin.utils.unquote(object_id)
+
+    def get_reversion_changeform_object_id(self, version):
+        obj = self.model._default_manager.using(version.db).get(slug=version.object_id)
+        return admin.utils.quote(str(obj.pk))
+
+
 class CustomObjectIdAdminMixin(TestBase):
 
     def setUp(self):
         super().setUp()
-        reversion.register(TestModelCustomObjectId, object_id_field="slug")
+        admin.site.register(TestModelCustomObjectId, TestModelCustomObjectIdAdmin)
         self.reloadUrls()
 
     def tearDown(self):
         super().tearDown()
+        admin.site.unregister(TestModelCustomObjectId)
         self.reloadUrls()
